requirements.txt
# Core Dependencies
streamlit>=1.28.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Computer Vision & AI
opencv-python>=4.8.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Visualization & 3D Graphics
plotly>=5.15.0
matplotlib>=3.7.0
networkx>=3.1.0
seaborn>=0.12.0

# Data Processing & Analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Space Science (Optional)
astropy>=5.3.0

# Utilities
schedule>=1.2.0
psutil>=5.9.0
python-dateutil>=2.8.0

# Development (Optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# API Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
slowapi==0.1.9
pydantic==2.5.0
aiohttp==3.9.0

# Computer Vision Dependencies
ultralytics==8.2.37
opencv-python-headless==4.10.0.84
torch==2.2.0
torchvision==0.17.0

# Time Series and AI Dependencies
tensorflow==2.15.0
transformers==4.35.0

# Solar Data Dependencies
drms
astropy
sunpy[map]
aiofiles
reproject